delete some TODOs that are no longer applicable
We don't need to verify `VM_METHOD_TYPE_SORBET` because callers have
already validated this for us.

We do still need to munge the send args, but the fastpath
implementations that this patch starts to implement can avoid those
mungings when possible.  So we need to add more of those fastpath
implementations.